userlist, info,update 세션처리

+++ ICT_MARKET/src/main/java/service/UserInfoDeleteService.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import kr.co.ict.UserDAO;
 
@@ -12,14 +13,17 @@ public class UserInfoDeleteService implements boardInterface_Service{
 
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
+		String user_id = request.getParameter("user_id");
 		String struNum = request.getParameter("user_num");
 		int uNum = Integer.parseInt(struNum);
-		
+		HttpSession session = request.getSession();
+		String sId = (String)session.getAttribute("session_id");
 		// 1. dao积己
+		if(sId == user_id) {
 		UserDAO dao = UserDAO.getInstance();
 		// 2. 昏力肺流 角青
 		dao.deleteUser(uNum);
-		
+		}
 	}
 
 }
